+++ test/amp_test.exs
@@ -1,27 +1,16 @@
 defmodule AmpTest do
   use ExUnit.Case
+  import Amp.TestHelpers, only: [
+    md_ext_to_html: 1,
+    remove_white_space: 1,
+    read_markdown: 1,
+    read_amp_html: 1,
+  ]
 
   test "Renders a h1 tag" do
     {:ok, html, []} = Amp.parse("# Hello World")
     assert html =~ "<h1>Hello World</h1>"
   end
-
-  defp html_ext(name), do: name <> ".html"
-
-  defp md_ext_to_html(markdown_file) do
-    markdown_file |> String.split(".") |> hd |> html_ext
-  end
-
-  defp remove_white_space(html) do
-    html
-    |> String.split("\n")
-    |> Enum.map(&(String.strip(&1)))
-    |> Enum.join("\n")
-  end
-
-  defp read(type, file), do: File.read("test/fixtures/#{type}/#{file}")
-  defp read_markdown(file), do: read("markdown", file)
-  defp read_amp_html(file), do: read("amp_html", file)
 
   test "Fixtures" do
     {:ok, markdown_files} = File.ls("test/fixtures/markdown")
